+++ packages/sp/batching.ts
@@ -147,7 +147,10 @@ export function createBatch(base: ISPQueryable, props?: ISPBatchProps): [Timelin
         await Promise.all(registrationPromises);
 
         if (requests.length < 1) {
-            return;
+            // even if we have no requests we need to await the complete promises to ensure
+            // that execute only resolves AFTER every child request disposes #2457
+            // this likely means caching is being used, we returned values for all child requests from the cache
+            return Promise.all(completePromises).then(() => void (0));
         }
 
         const batchBody: string[] = [];
@@ -369,8 +372,10 @@ function parseResponse(body: string): Response[] {
     let status: number;
     let statusText: string;
     let headers = {};
+    const bodyReader = [];
+
     for (let i = 0; i < lines.length; ++i) {
-        const line = lines[i];
+        let line = lines[i];
         switch (state) {
             case "batch":
                 if (line.substring(0, header.length) === header) {
@@ -407,7 +412,19 @@ function parseResponse(body: string): Response[] {
                 }
                 break;
             case "body":
-                responses.push(new Response(status === 204 ? null : line, { status, statusText, headers }));
+
+                // reset the body reader
+                bodyReader.length = 0;
+                // this allows us to capture batch bodies that are returned as multi-line (renderListDataAsStream, #2454)
+                while (line.substring(0, header.length) !== header) {
+                    bodyReader.push(line);
+                    line = lines[++i];
+                }
+                // because we have read the closing --batchresponse_ line, we need to move the line pointer back one
+                // so that the logic works as expected either to get the next result or end processing
+                i--;
+
+                responses.push(new Response(status === 204 ? null : bodyReader.join(""), { status, statusText, headers }));
                 state = "batch";
                 headers = {};
                 break;