+++ test/graph/sites.ts
@@ -1,7 +1,10 @@
 import { expect } from "chai";
+import getTestingGraphSPSite from "./utilities/getTestingGraphSPSite.js";
 import "@pnp/graph/sites";
+import { ISite } from "@pnp/graph/sites";
 
 describe("Sites", function () {
+    let site: ISite;
 
     before(async function () {
 
@@ -19,6 +22,16 @@ describe("Sites", function () {
         let passed = true;
         const site = await this.pnp.graph.sites.getById(this.pnp.settings.graph.id)();
         passed = (site.id === this.pnp.settings.graph.id);
+        return expect(passed).is.true;
+    });
+
+    it("getByUrl()", async function() {
+        let passed = true;
+        site = await getTestingGraphSPSite(this);
+        const tetssite = await site();
+        const url = new URL(tetssite.webUrl);
+        const siteByUrl = await this.pnp.graph.sites.getByUrl(url.hostname, url.pathname)();
+        passed = (siteByUrl.webUrl.toLowerCase() === tetssite.webUrl.toLowerCase());
         return expect(passed).is.true;
     });