+++ test/mocha-root-hooks.ts
@@ -113,15 +113,7 @@ export const mochaHooks = {
                 this.pnp.settings.sp.testWebUrl = testWebResult.data.Url;
 
                 // create a new testing site
-                this.pnp._sp = spfi(this.pnp.settings.sp.testWebUrl).using(
-                    SPDefault({
-                        msal: {
-                            config: this.pnp.settings.sp.msal.init,
-                            scopes: this.pnp.settings.sp.msal.scopes,
-                        },
-                    }),
-                    NodeFetch({ replace: true }),
-                    PnPLogging(this.pnp.args.logging));
+                this.pnp._sp = spfi([rootSP.web, this.pnp.settings.sp.testWebUrl]);
 
                 // TODO:: remove once pnpTest is used everywhere
                 this.pnp.sp = this.pnp._sp;